// src/shared/api/models/getApiStatisticsSalesSummary200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiStatisticsSalesSummary200Summary } from './getApiStatisticsSalesSummary200Summary';
import type { GetApiStatisticsSalesSummary200SoldItemsItem } from './getApiStatisticsSalesSummary200SoldItemsItem';
import type { GetApiStatisticsSalesSummary200PaymentMethodItem } from './getApiStatisticsSalesSummary200PaymentMethodItem';

/**
 * Successfully retrieved sales summary statistics
 */
export type GetApiStatisticsSalesSummary200 = {
    summary: GetApiStatisticsSalesSummary200Summary;
    soldItems: GetApiStatisticsSalesSummary200SoldItemsItem[];
    paymentMethod: GetApiStatisticsSalesSummary200PaymentMethodItem[];
};